+++ src/string.h
@@ -5,8 +5,8 @@
 #include "error.h"
 
 typedef struct String {
-	size_t count_bytes;
-	size_t capacity_bytes;
+	uint64_t count_bytes;
+	uint64_t capacity_bytes;
 	char* text;
 } String;
 
@@ -19,7 +19,7 @@ String_Result
 string_new(const char* cstr);
 
 String_Result
-string_new_with_count_bytes(size_t cstr_count_bytes, const char* cstr);
+string_new_with_count_bytes(uint64_t cstr_count_bytes, const char* cstr);
 
 void
 string_free(String* s);
@@ -31,9 +31,9 @@ String_Result
 string_append_cstr(String* a, const char* b);
 
 String_Result
-string_append_cstr_with_count_bytes(String* a, size_t b_count_bytes, const char* b);
+string_append_cstr_with_count_bytes(String* a, uint64_t b_count_bytes, const char* b);
 
 Error
-string_reserve(String* s, size_t n);
+string_reserve(String* s, uint64_t n);
 
 #endif // STRING_H
